Atualização do código UserService-DELETE

+++ src/main/java/com/educandoweb/course/services/UserService.java
@@ -38,14 +38,17 @@ public class UserService {
 	}
 
 	public void delete(Long id) {
+		
 		try {
-			repository.deleteById(id);
-		} catch (EmptyResultDataAccessException e) {
-			throw new ResourceNotFoundException(id);
+	        if(!repository.existsById(id)) throw new ResourceNotFoundException(id);
+	        repository.deleteById(id);
+	        
+	    } catch (ResourceNotFoundException e){
+	        throw new ResourceNotFoundException(id);
+	    
 		} catch (DataIntegrityViolationException e) {
 			throw new DatabaseException(e.getMessage());
 		}
-
 	}
 
 	public User update(Long id, User obj) {
